code/requirements.txt
python-dotenv~=0.20.0
tweepy~=4.8.0
pandas~=1.4.2
wordcloud~=1.8.1
matplotlib~=3.5.1
nltk~=3.7
Unidecode~=1.3.4
scikit-learn~=1.0.2